+++ alex/applications/PublicTransportInfoCS/data/expand_stops.py
@@ -25,52 +25,11 @@ import re
 from collections import defaultdict
 import itertools
 import codecs
-from ufal.morphodita import Tagger, Forms, TaggedLemmas, TokenRanges, Morpho, TaggedLemmasForms
 
 from alex.utils.config import online_update
 from alex.utils.various import remove_dups_stable
+from alex.applications.PublicTransportInfoCS.cs_morpho import Analyzer, Generator 
 from getopt import getopt
-
-
-class Analyzer(object):
-    """Morphodita analyzer/tagger wrapper."""
-
-    def __init__(self, tagger_model):
-        self.__tagger = Tagger.load(tagger_model)
-        self.__tokenizer = self.__tagger.newTokenizer()
-        self.__forms_buf = Forms()
-        self.__tokens_buf = TokenRanges()
-        self.__lemmas_buf = TaggedLemmas()
-
-    def analyze(self, stop_text):
-        self.__tokenizer.setText(stop_text)
-        while self.__tokenizer.nextSentence(self.__forms_buf, self.__tokens_buf):
-            self.__tagger.tag(self.__forms_buf, self.__lemmas_buf)
-            return [(form, lemma.lemma, lemma.tag)
-                    for (form, lemma) in zip(self.__forms_buf, self.__lemmas_buf)]
-
-
-class Generator(object):
-    """Morphodita generator wrapper."""
-
-    def __init__(self, morpho_model):
-        self.__morpho = Morpho.load(morpho_model)
-        self.__out_buf = TaggedLemmasForms()
-
-    def generate(self, lemma, tag_wildcard, capitalized=None):
-        # run the generation for this word
-        self.__morpho.generate(lemma, tag_wildcard, self.__morpho.GUESSER, self.__out_buf)
-        # see if we found any forms, return empty if not
-        if not self.__out_buf:
-            return []
-        # prepare capitalization
-        cap_func = lambda string: string
-        if capitalized == True:
-            cap_func = lambda string: string[0].upper() + string[1:]
-        elif capitalized == False:
-            cap_func = lambda string: string[0].lower() + string[1:]
-        # process the results
-        return [cap_func(form_tag.form) for form_tag in self.__out_buf[0].forms]
 
 
 class ExpandStops(object):
@@ -149,11 +108,7 @@ class ExpandStops(object):
                     words = self.__analyzer.analyze(variant)
                     # in all required cases
                     for case in self.cases_list:
-                        forms = []
-                        prev_tag = ''
-                        for word in words:
-                            forms.append(self.__inflect_word(word, case, prev_tag))
-                            prev_tag = word[2]
+                        forms = self.__generator.inflect(words, case, self.personal_names)
                         # use all possible combinations if there are more variants for this case
                         inflected = map(self.__postprocess_func,
                                         remove_dups_stable([' '.join(var)
@@ -163,28 +118,6 @@ class ExpandStops(object):
                 if ctr % 1000 == 0:
                     print >> sys.stderr, '.',
         print >> sys.stderr
-
-    def __inflect_word(self, word, case, prev_tag):
-        """Inflect one word in the given case (return a list of variants)."""
-        form, lemma, tag = word
-        # inflect each word in nominative not following a noun in nominative
-        # (if current case is not nominative), avoid numbers
-        if (re.match(r'^[^C]...1', tag) and
-                (not re.match(r'^NN..1', prev_tag) or self.personal_names) and
-                form not in ['římská'] and
-                case != '1'):
-            # change the case in the tag, allow all variants
-            new_tag = re.sub(r'^(....)1(.*).$', r'\g<1>' + case + r'\g<2>?', tag)
-            # -ice: test both sg. and pl. versions
-            if (form.endswith('ice') and form[0] == form[0].upper() and
-                    not re.match(r'(nemocnice|ulice|vrátnice)', form, re.IGNORECASE)):
-                new_tag = re.sub(r'^(...)S', r'\1[SP]', new_tag)
-            # try inflecting, fallback to uninflected
-            capitalized = form[0] == form[0].upper()
-            new_forms = self.__generator.generate(lemma, new_tag, capitalized)
-            return new_forms if new_forms else [form]
-        else:
-            return [form]
 
 
 def main():
